requirements.txt
groq
requests
pydantic
beautifulsoup4
python-dotenv
typing-extensions
pytest
urllib3
requests-html
flask
chardet
#dateutil
firecrawl-py
groq
selenium


gspread
oauth2client

google-auth-oauthlib
google-auth-httplib2
google-api-python-client

tldextract
google-generativeai
google-genai
pymupdf
openpyxl
